+++ SuperDUPER/Font.cpp
@@ -2,7 +2,7 @@
 
 
 
-Font::Font(short fontSize, const std::string& fontFileName)
+Font::Font(int fontSize, const std::string& fontFileName)
 {
 	font = TTF_OpenFont(fontFileName.c_str(), static_cast<int>(fontSize));
 	if (!font)
