+++ SuperDUPER/Item.h
@@ -3,12 +3,12 @@
 class Item
 {
 public:
-	Item(const std::string& name, short mass);
-	void setMass(short newMass) { mass = newMass; }
-	short getMass() const { return mass; }
+	Item(const std::string& name, int mass);
+	void setMass(int newMass) { mass = newMass; }
+	int getMass() const { return mass; }
 	~Item();
 private:
 	std::string name;
-	short mass;
+	int mass;
 };
