+++ SuperDUPER/WeaponAttack.cpp
@@ -2,7 +2,7 @@
 
 
 
-WeaponAttack::WeaponAttack(short range, float damageMultiplier, float sharpness)
+WeaponAttack::WeaponAttack(int range, float damageMultiplier, float sharpness)
 	:range{ range }, damageMultiplier{ damageMultiplier }, sharpness{ sharpness }
 {
 }
